backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
neomodel==5.3.2
django-neomodel==0.2.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
